Why the compiler doesn't catch this? Should be double instead of int #397

+++ lib/devices/device_descriptors/kaya_first_descriptor.dart
@@ -61,7 +61,7 @@ class KayakFirstDescriptor extends DeviceDescriptor {
     final dataString = utf8.decode(data);
     final dataParts = dataString.split(";");
     return RecordWithSport(
-      distance: int.tryParse(dataParts[9]),
+      distance: double.tryParse(dataParts[9]),
       elapsed: int.tryParse(dataParts[22]),
       power: int.tryParse(dataParts[21]),
       speed: double.tryParse(dataParts[11]) ?? 0.0 * DeviceDescriptor.ms2kmh,
